# specs/api.yaml
openapi: 3.1.0
info:
  title: PsiZero Resonance Platform API
  version: 1.0.0
  description: >
    Integrated APIs for symbolic resonance technologies:
    - SRS (Symbolic Resonance Solver) for NP-complete problems
    - HQE (Holographic Quantum Encoder) simulator
    - QSEM (Quantum Semantics) for concept resonance
    - NLC (Non-Local Communication) prime-eigenstate channels
    - QCR (Quantum Consciousness Resonator) sessions
    - I-CHING oracle (hexagram evolution)
    - Unified resonance physics (observer-capacity/gravity model)
servers:
  - url: https://api.psizero.dev
    description: Production
  - url: https://sandbox.psizero.dev
    description: Sandbox
...
security:
  - ApiKeyAuth: []
  - OAuth2CC: [platform.read, platform.write]
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    OAuth2CC:
      type: oauth2
      description: Client Credentials flow (for enterprises)
      flows:
        clientCredentials:
          tokenUrl: https://auth.psizero.dev/oauth/token
          scopes:
            platform.read: Read access
            platform.write: Write/access compute
...
  responses:
    ProblemJson:
      description: RFC 7807 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string }
        meta: { type: object, additionalProperties: true }
...
    ResonanceMetrics:
      type: object
      properties:
        entropy: { type: number }
        plateauDetected: { type: boolean }
        dominance: { type: number, description: "max mass / coherence, [0,1]" }
        locality: { type: number, description: "non-locality inverse; lower=more non-local" }
        resonanceStrength: { type: number, description: "phase-lock strength, [0,1]" }

    # ---------- SRS ----------
...
    HQERequest:
      type: object
      properties:
        primes:
          type: array
          description: Prime eigenstates to include
          items: { type: integer, minimum: 2 }
        dt: { type: number, default: 0.1 }
        steps: { type: integer, default: 256 }
        lambda: { type: number, default: 0.02, description: entropy dissipation }
        resonanceTarget: { type: number, description: r_stable }
        initialAmplitudes:
          type: array
          items: { type: number }
      required: [primes]
    HQESnapshot:
      type: object
      properties:
        step: { type: integer }
        amplitudes:
          type: array
          items: { type: number }
        metrics: { $ref: '#/components/schemas/ResonanceMetrics' }
    HQEResponse:
      type: object
      properties:
        snapshots:
          type: array
          items: { $ref: '#/components/schemas/HQESnapshot' }
        finalMetrics: { $ref: '#/components/schemas/ResonanceMetrics' }

    # ---------- QSEM ----------
...
paths:
...
  /v1/hqe/simulate:
    post:
      tags: [HQE]
      summary: Run a Holographic Quantum Encoder simulation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HQERequest' }
            examples:
              basic:
                value: { primes: [2,3,5,7], steps: 128, lambda: 0.03, resonanceTarget: 0.8 }
      responses:
        '200':
          description: Time-series snapshots and final metrics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HQEResponse' }
        '400': { $ref: '#/components/responses/ProblemJson' }

  # ---------- QSEM ----------
...